# app/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the Razee.io API server.  You can find out more about Razee at [http://razee.io](http://razee.io) or on [slack, #razee](https://ibm-container-service.slack.com/messages/CJQG07QTX).      For this sample, you can use the api key `special-key` to test the authorization     filters.",
    "version": "2.0.0",
    "title": "Razeedash API",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3333",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/channels": {
      "post": {
        "tags": [
          "channels"
        ],
        "summary": "Adds a new channel",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      },
      "get": {
        "tags": [
          "channels"
        ],
        "summary": "Gets all channels for an org",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/channels/{uuid}/versions": {
      "post": {
        "tags": [
          "channels"
        ],
        "summary": "Adds a new channel version",
        "description": "",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "uuid": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "type": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/clusters/clustersByOrgId": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "Gets all clusters for an org",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/groups": {
      "post": {
        "tags": [
          "groups"
        ],
        "summary": "Adds a group",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/groups/groupClusters": {
      "put": {
        "tags": [
          "groups"
        ],
        "summary": "Assigns a group to a cluster",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "uuid": {
                  "example": "any"
                },
                "clusters": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/subscriptions": {
      "post": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Adds a subscription",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "groups": {
                  "example": "any"
                },
                "clusterId": {
                  "example": "any"
                },
                "channelUuid": {
                  "example": "any"
                },
                "versionUuid": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/subscriptions/{uuid}": {
      "get": {
        "tags": [
          "subscriptions"
        ],
        "summary": "Gets a subscription",
        "description": "",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    }
  },
  "securityDefinitions": {
    "razee-org-key": {
      "type": "apiKey",
      "name": "razee-org-key",
      "in": "header"
    },
    "org-admin-key": {
      "type": "apiKey",
      "name": "org-admin-key",
      "in": "header"
    }
  },
...
}